requirements.txt
matplotlib
numpy
pandas
cairosvg
scikit-learn>=0.20.2
biopython